fix jacoco reports and integration tests enabling
Jacoco Reports:
in a multimodule project, you need the jacoco-report-aggregation plugin
to aggregate the jacoco reports of all the subprojects. I created a new
subproject where you can find the reports aggregated in a single file

Integration tests enabling:
the system property passed to 'gradle test' must be propagated to the tests
manually, see the addition in the test task in build.gradle. Then i have
added the possibility to start the workflow manually and specify if the
integration tests should be enabled