add test for shop subscription and message publishing
wip: how to retrieve the published message? is there an internal localstack endpoint?
the getInfo endpoint breaks because now the idToken is invalid and the error message is
An error occurred (UserNotFoundException) when calling the AdminGetUser operation: User does not exist.
check if the selection pattern is usable or not